subMacroMultiCheckConn.sh: init
Implement a simple connection check to each input file URL in each
input file list.

This scripts check the connection by performing
`root -l -q $inputFileUrl`, and collecting the output.
If the keyword "nullptr" is found in the output, it indicates a
connection failure. The `$inputFileUrl`s failed to connect will be
logged out to the standard output.

During the running, the stdout and stderr of the test command is also
piped to stderr.